Migrate all Controllers from ASP.NET Framework to ASP.NET Core MVC
- HomeController: 已更新為 ASP.NET Core MVC，加入建構子注入 SchoolContext 和 INotificationService，About 方法改為非同步
- StudentsController: 所有 action 方法改為非同步，更新 Bind 屬性語法，改用 BadRequest() 和 NotFound() 回傳
- CoursesController: 支援非同步檔案上傳，使用 IWebHostEnvironment 和 IFormFile 取代 HttpPostedFileBase
- InstructorsController: 所有資料庫操作改為非同步，更新 TryUpdateModelAsync 語法為表達式樹
- DepartmentsController: 所有 action 方法改為非同步，處理並行衝突的邏輯改用 async/await
- ToDosController: 使用非同步 ADO.NET 操作，所有 stored procedure 調用改為非同步
- NotificationsController: 更新為使用 JsonResult 而非舊的 Json() 方法
- PaginatedList: 新增 CreateAsync 方法支援非同步分頁
- SchoolContextFactory: 更新使用 IConfiguration 取代 ConfigurationManager
- Views/_ViewImports.cshtml: 新增以支援 Razor 視圖中的命名空間
- Views/Shared/Error.cshtml: 移除 System.Web.Mvc.HandleErrorInfo 依賴

所有 Controller 現在完全相容 ASP.NET Core MVC，專案編譯成功。

+++ ContosoUniversity/Controllers/InstructorsController.cs
@@ -1,27 +1,33 @@
 using System;
 using System.Collections.Generic;
+using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using System.Linq;
-using System.Net;
-using System.Web.Mvc;
+using Microsoft.AspNetCore.Mvc;
 using ContosoUniversity.Data;
 using ContosoUniversity.Models;
 using ContosoUniversity.Models.SchoolViewModels;
+using ContosoUniversity.Services;
 
 namespace ContosoUniversity.Controllers
 {
     public class InstructorsController : BaseController
     {
+        public InstructorsController(SchoolContext context, INotificationService notification) 
+            : base(context, notification)
+        {
+        }
         // GET: Instructors - All roles can view
-        public ActionResult Index(int? id, int? courseID)
+        public async Task<IActionResult> Index(int? id, int? courseID)
         {
             var viewModel = new InstructorIndexData();
-            viewModel.Instructors = db.Instructors
+            viewModel.Instructors = await db.Instructors
                 .Include(i => i.OfficeAssignment)
                 .Include(i => i.CourseAssignments)
                     .ThenInclude(c => c.Course)
                         .ThenInclude(d => d.Department)
-                .OrderBy(i => i.LastName);
+                .OrderBy(i => i.LastName)
+                .ToListAsync();
 
             if (id != null)
             {
@@ -41,33 +47,33 @@ namespace ContosoUniversity.Controllers
         }
 
         // GET: Instructors/Details/5 - All roles can view details
-        public ActionResult Details(int? id)
+        public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return BadRequest();
             }
-            Instructor instructor = db.Instructors.Find(id);
+            Instructor? instructor = await db.Instructors.FindAsync(id);
             if (instructor == null)
             {
-                return HttpNotFound();
+                return NotFound();
             }
             return View(instructor);
         }
 
         // GET: Instructors/Create
-        public ActionResult Create()
+        public async Task<IActionResult> Create()
         {
             var instructor = new Instructor();
             instructor.CourseAssignments = new List<CourseAssignment>();
-            PopulateAssignedCourseData(instructor);
+            await PopulateAssignedCourseDataAsync(instructor);
             return View(instructor);
         }
 
         // POST: Instructors/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "LastName,FirstMidName,HireDate,OfficeAssignment")] Instructor instructor, string[] selectedCourses)
+        public async Task<IActionResult> Create([Bind("LastName,FirstMidName,HireDate,OfficeAssignment")] Instructor instructor, string[]? selectedCourses)
         {
             if (selectedCourses != null)
             {
@@ -81,41 +87,41 @@ namespace ContosoUniversity.Controllers
             if (ModelState.IsValid)
             {
                 db.Instructors.Add(instructor);
-                db.SaveChanges();
+                await db.SaveChangesAsync();
                 
                 // Send notification for instructor creation
                 SendEntityNotification("Instructor", instructor.ID.ToString(), EntityOperation.CREATE);
                 
                 return RedirectToAction("Index");
             }
-            PopulateAssignedCourseData(instructor);
+            await PopulateAssignedCourseDataAsync(instructor);
             return View(instructor);
         }
 
         // GET: Instructors/Edit/5
-        public ActionResult Edit(int? id)
+        public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return BadRequest();
             }
-            Instructor instructor = db.Instructors
+            Instructor? instructor = await db.Instructors
                 .Include(i => i.OfficeAssignment)
                 .Include(i => i.CourseAssignments)
                     .ThenInclude(c => c.Course)
                 .Where(i => i.ID == id)
-                .Single();
-            PopulateAssignedCourseData(instructor);
+                .FirstOrDefaultAsync();
             if (instructor == null)
             {
-                return HttpNotFound();
+                return NotFound();
             }
+            await PopulateAssignedCourseDataAsync(instructor);
             return View(instructor);
         }
 
-        private void PopulateAssignedCourseData(Instructor instructor)
+        private async Task PopulateAssignedCourseDataAsync(Instructor instructor)
         {
-            var allCourses = db.Courses;
+            var allCourses = await db.Courses.ToListAsync();
             var instructorCourses = new HashSet<int>(instructor.CourseAssignments.Select(c => c.CourseID));
             var viewModel = new List<AssignedCourseData>();
             foreach (var course in allCourses)
@@ -133,32 +139,37 @@ namespace ContosoUniversity.Controllers
         // POST: Instructors/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit(int? id, string[] selectedCourses)
+        public async Task<IActionResult> Edit(int? id, string[]? selectedCourses)
         {
             if (id == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return BadRequest();
             }
-            var instructorToUpdate = db.Instructors
+            var instructorToUpdate = await db.Instructors
                .Include(i => i.OfficeAssignment)
                .Include(i => i.CourseAssignments)
                    .ThenInclude(c => c.Course)
                .Where(i => i.ID == id)
-               .Single();
+               .FirstOrDefaultAsync();
 
-            if (TryUpdateModel(instructorToUpdate, "",
-               new string[] { "LastName", "FirstMidName", "HireDate", "OfficeAssignment" }))
+            if (instructorToUpdate == null)
+            {
+                return NotFound();
+            }
+
+            if (await TryUpdateModelAsync(instructorToUpdate, "",
+               i => i.LastName, i => i.FirstMidName, i => i.HireDate, i => i.OfficeAssignment))
             {
                 try
                 {
-                    if (String.IsNullOrWhiteSpace(instructorToUpdate.OfficeAssignment.Location))
+                    if (instructorToUpdate.OfficeAssignment != null && String.IsNullOrWhiteSpace(instructorToUpdate.OfficeAssignment.Location))
                     {
                         instructorToUpdate.OfficeAssignment = null;
                     }
 
-                    UpdateInstructorCourses(selectedCourses, instructorToUpdate);
+                    await UpdateInstructorCoursesAsync(selectedCourses, instructorToUpdate);
 
-                    db.SaveChanges();
+                    await db.SaveChangesAsync();
                     
                     // Send notification for instructor update
                     SendEntityNotification("Instructor", instructorToUpdate.ID.ToString(), EntityOperation.UPDATE);
@@ -170,11 +181,11 @@ namespace ContosoUniversity.Controllers
                     ModelState.AddModelError("", "Unable to save changes. Try again, and if the problem persists, see your system administrator.");
                 }
             }
-            PopulateAssignedCourseData(instructorToUpdate);
+            await PopulateAssignedCourseDataAsync(instructorToUpdate);
             return View(instructorToUpdate);
         }
 
-        private void UpdateInstructorCourses(string[] selectedCourses, Instructor instructorToUpdate)
+        private async Task UpdateInstructorCoursesAsync(string[]? selectedCourses, Instructor instructorToUpdate)
         {
             if (selectedCourses == null)
             {
@@ -185,7 +196,8 @@ namespace ContosoUniversity.Controllers
             var selectedCoursesHS = new HashSet<string>(selectedCourses);
             var instructorCourses = new HashSet<int>
                 (instructorToUpdate.CourseAssignments.Select(c => c.Course.CourseID));
-            foreach (var course in db.Courses)
+            var allCourses = await db.Courses.ToListAsync();
+            foreach (var course in allCourses)
             {
                 if (selectedCoursesHS.Contains(course.CourseID.ToString()))
                 {
@@ -199,24 +211,27 @@ namespace ContosoUniversity.Controllers
 
                     if (instructorCourses.Contains(course.CourseID))
                     {
-                        CourseAssignment courseToRemove = instructorToUpdate.CourseAssignments.SingleOrDefault(i => i.CourseID == course.CourseID);
-                        db.Entry(courseToRemove).State = EntityState.Deleted;
+                        CourseAssignment? courseToRemove = instructorToUpdate.CourseAssignments.SingleOrDefault(i => i.CourseID == course.CourseID);
+                        if (courseToRemove != null)
+                        {
+                            db.Entry(courseToRemove).State = EntityState.Deleted;
+                        }
                     }
                 }
             }
         }
 
         // GET: Instructors/Delete/5
-        public ActionResult Delete(int? id)
+        public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return BadRequest();
             }
-            Instructor instructor = db.Instructors.Find(id);
+            Instructor? instructor = await db.Instructors.FindAsync(id);
             if (instructor == null)
             {
-                return HttpNotFound();
+                return NotFound();
             }
             return View(instructor);
         }
@@ -224,38 +239,34 @@ namespace ContosoUniversity.Controllers
         // POST: Instructors/Delete/5 - Only admins can delete instructors
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            Instructor instructor = db.Instructors
+            Instructor? instructor = await db.Instructors
               .Include(i => i.OfficeAssignment)
               .Where(i => i.ID == id)
-              .Single();
+              .FirstOrDefaultAsync();
+
+            if (instructor == null)
+            {
+                return NotFound();
+            }
 
             db.Instructors.Remove(instructor);
 
-            var department = db.Departments
+            var department = await db.Departments
                 .Where(d => d.InstructorID == id)
-                .SingleOrDefault();
+                .FirstOrDefaultAsync();
             if (department != null)
             {
                 department.InstructorID = null;
             }
 
-            db.SaveChanges();
+            await db.SaveChangesAsync();
             
             // Send notification for instructor deletion
             SendEntityNotification("Instructor", id.ToString(), EntityOperation.DELETE);
             
             return RedirectToAction("Index");
-        }
-
-        protected override void Dispose(bool disposing)
-        {
-            if (disposing)
-            {
-                db.Dispose();
-            }
-            base.Dispose(disposing);
         }
     }
 }
